add max_slider_value to config
# Notes
For sliders that cannot reach the full analog range, use max_slider_value in the config.yaml to adjust the range from 0 to max_slider_value.

I wrote this because I made the mistake of using Logarithmic Potentiometers, which only reach a value of 900.

Converting a logarithmic value to linear, wasn't ideal for range / sensitivity.

So adjusting the max value in range, works very well to remedy this issue.